Fix types-related warnings based on clang diagnostics
1. Fix some places where unsigned value compared < 0
2. Fix obsolete %Z specifier to more portable %z
3. Some erroneous types substitution
4. nl_msec2str() - 64-bit msec is now properly used,

Only safe changes. I mean int <--> uint32_t and signed/unsigned fixes.
Some functinos require size_t argument instead of int, but changes of
signatures of that functions is terrible thing.

Also, I do not pretend for a full list of fixes.
Just to shut up clang -Wall -Wextra

One more thing. ifindex. I don't change that because changes will
be too big for simple fix.

+++ lib/addr.c
@@ -368,7 +368,7 @@ int nl_addr_parse(const char *addrstr, int hint, struct nl_addr **result)
 	}
 
 	if (hint == AF_UNSPEC && strchr(str, ':')) {
-		int i = 0;
+		size_t i = 0;
 		char *s = str, *p;
 		for (;;) {
 			long l = strtol(s, &p, 16);
@@ -542,7 +542,7 @@ int nl_addr_cmp_prefix(struct nl_addr *a, struct nl_addr *b)
  */
 int nl_addr_iszero(struct nl_addr *addr)
 {
-	int i;
+	unsigned int i;
 
 	for (i = 0; i < addr->a_len; i++)
 		if (addr->a_addr[i])
@@ -823,7 +823,7 @@ unsigned int nl_addr_get_prefixlen(struct nl_addr *addr)
  */
 char *nl_addr2str(struct nl_addr *addr, char *buf, size_t size)
 {
-	int i;
+	unsigned int i;
 	char tmp[16];
 
 	if (!addr || !addr->a_len) {
